Renamed binding.cpp to blake2.cpp
